requirements.txt
pydantic>=2.9
PyYAML>=6.0
google-genai>=1.68.0
google-generativeai
PyMuPDF>=1.24.0
python-dotenv
